cleanup, add entrypoints to sklearn and pt deployments
