+++ sage/cli.py
@@ -122,7 +122,7 @@ def deploy(
     elif image_uri == "sklearn":
         sklearn_model = SKLearnModel(
             model_data=model_path,
-            entry_point="",  # fill in
+            entry_point=entry_point,  # fill in
             role=role,
             framework_version="1.2-1",
             py_version="py3",
@@ -136,7 +136,7 @@ def deploy(
     elif image_uri == "pytorch":
         pytorch_model = PyTorchModel(
             model_data=model_path,
-            entry_point="",  #  fill in
+            entry_point=entry_point,  #  fill in
             role=role,
             framework_version="2.0.0",
             py_version="py310",
