--- sage/example_entrypoints/pt_dummy.py
@@ -1,10 +1,0 @@
-import torch.nn as nn
-
-
-class Dummy(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.fc1 = nn.Linear(1, 1)
-
-    def forward(self, x):
-        return 1
