+++ sage/example_entrypoints/pytorch_dummy_entrypoint.py
@@ -1,7 +1,16 @@
 import os
 import torch
 import json
-from pt_dummy import Dummy
+import torch.nn as nn
+
+
+class Dummy(nn.Module):
+    def __init__(self):
+        super().__init__()
+        self.fc1 = nn.Linear(1, 1)
+
+    def forward(self, x):
+        return 1
 
 
 def model_fn(model_path: str):
