--- sage/example_entrypoints/save_dummy.py
@@ -1,6 +1,0 @@
-import torch
-
-from pt_dummy import Dummy
-
-dummy_model = Dummy()
-torch.save(dummy_model.state_dict(), "dummy.pt")
